Refactor workflow and build scripts for different OS
The commit separates the tasks of the build script for Linux and Windows environments in the integration workflow. It also removes unused imports in the build.cs file. The 'CreateWin64Setup' task is now restricted to only run when the OS environment is Windows, while GitHub Actions will separately run 'deploynuget' for both environments.