Update workflows to publish setup artifacts
In the workflows 'release.yml', 'integration-win.yml', and 'main.yml', new steps have been added to publish the 'setups' artifact. This change allows the storage of generated setups for potential uses in future steps or later stages of the workflows. For more granular control, a generalized file name filter was also added in the 'Build' class.