Add GitHub Actions for Windows integration and update builds
This commit does several updates in the application's GitHub actions and build settings. It includes a new GitHub Actions workflow to handle 'feature/**' branches on Windows. Additionally, the build commands for the main workflow and release workflow have been modified. Lastly, support for creating a GitHub Release with the setup artefacts is added.