in creating GithubReleaseAsset
The revised code now uses the complete filename in the creation of a GithubReleaseAsset, rather than just obtaining the file name. This adjustment has also been carried out so as to make a null Stream, instead of opening the file to read it. This should improve the efficiency of the code by reducing the need for file operations.