Create Github release on release build with Windows setup for gt (#8)
* Add GitHub Actions for Windows integration and update builds

This commit does several updates in the application's GitHub actions and build settings. It includes a new GitHub Actions workflow to handle 'feature/**' branches on Windows. Additionally, the build commands for the main workflow and release workflow have been modified. Lastly, support for creating a GitHub Release with the setup artefacts is added.

* Add pre-release task to Windows setup creation

A pre-release task was added to the Windows setup creation process. This task ensures that the necessary actions and checks are performed before creating a GitHub release. This should improve the reliability and robustness of the release process.

* in creating GithubReleaseAsset

The revised code now uses the complete filename in the creation of a GithubReleaseAsset, rather than just obtaining the file name. This adjustment has also been carried out so as to make a null Stream, instead of opening the file to read it. This should improve the efficiency of the code by reducing the need for file operations.

* Update invoked targets in GitHub actions

In the build configuration file, the list of invoked targets for GitHub actions has been revised. Instead of "deploynuget", "rebuild" and "codecoverage" are now included in the target actions to better reflect the requirements of the current pipeline.

* Update GitHub Actions workflow commands

The commands in the GitHub Actions workflow script for Windows integration have been updated. The 'deploynuget' task has been replaced with a 'rebuild' task, and a new 'codecoverage' task was added to ensure that code coverage is calculated during the build process.

* Update build scripts and package versions

Capitalized the build commands used in Github Actions workflows. Also changed the package version of "CreativeCoders.Core" from 4.1.0 to 4.2.0 in multiple project files and updated other related package versions in build scripts for consistency.

* Remove integration-win workflow

The integration-win.yml workflow file and its related configuration in Build.cs have been deleted. This workflow was designed to run specific operations on each push to feature branches running on windows-latest. It is no longer required and therefore has been removed.